Add hook for ItemViewerContext and move types to a different file
