+++ content/webapp/components/IIIFViewer/GridViewer.tsx
@@ -3,7 +3,6 @@ import {
   CSSProperties,
   FunctionComponent,
   memo,
-  useContext,
   useEffect,
   useRef,
   useState,
@@ -15,11 +14,10 @@ import { useAppContext } from '@weco/common/contexts/AppContext';
 import LL from '@weco/common/views/components/styled/LL';
 import Space from '@weco/common/views/components/styled/Space';
 import { toLink as itemLink } from '@weco/content/components/ItemLink';
-import ItemViewerContext, {
-  Query,
-} from '@weco/content/contexts/ItemViewerContext';
+import { useItemViewerContext } from '@weco/content/contexts/ItemViewerContext';
 import useScrollVelocity from '@weco/content/hooks/useScrollVelocity';
 import { SearchResults } from '@weco/content/services/iiif/types/search/v3';
+import { ItemViewerQuery } from '@weco/content/types/item-viewer';
 import { TransformedCanvas } from '@weco/content/types/manifest';
 
 import { arrayIndexToQueryParam, queryParamToArrayIndex } from '.';
@@ -49,7 +47,7 @@ type CellProps = {
     setGridVisible: (value: boolean) => void;
     canvases: TransformedCanvas[];
     searchResults: SearchResults;
-    query: Query;
+    query: ItemViewerQuery;
     workId: string;
   };
 };
@@ -139,7 +137,7 @@ const GridViewer: FunctionComponent = () => {
     searchResults,
     query,
     work,
-  } = useContext(ItemViewerContext);
+  } = useItemViewerContext();
   const { windowSize } = useAppContext();
   const [newScrollOffset, setNewScrollOffset] = useState(0);
   const scrollVelocity = useScrollVelocity(newScrollOffset);
