+++ content/webapp/components/IIIFViewer/MainViewer.tsx
@@ -4,7 +4,6 @@ import {
   CSSProperties,
   FunctionComponent,
   memo,
-  useContext,
   useEffect,
   useRef,
   useState,
@@ -16,11 +15,10 @@ import { useUserContext } from '@weco/common/contexts/UserContext';
 import { font } from '@weco/common/utils/classnames';
 import LL from '@weco/common/views/components/styled/LL';
 import IIIFItem from '@weco/content/components/IIIFItem';
-import ItemViewerContext, {
-  RotatedImage,
-} from '@weco/content/contexts/ItemViewerContext';
+import { useItemViewerContext } from '@weco/content/contexts/ItemViewerContext';
 import useScrollVelocity from '@weco/content/hooks/useScrollVelocity';
 import { SearchResults } from '@weco/content/services/iiif/types/search/v3';
+import { CanvasRotatedImage } from '@weco/content/types/item-viewer';
 import { TransformedCanvas } from '@weco/content/types/manifest';
 import { TransformedAuthService } from '@weco/content/utils/iiif/v3';
 import { getDisplayItems } from '@weco/content/utils/iiif/v3/canvas';
@@ -120,7 +118,7 @@ type ItemRendererProps = {
   data: {
     scrollVelocity: number;
     canvases: TransformedCanvas[];
-    rotatedImages: RotatedImage[];
+    rotatedImages: CanvasRotatedImage[];
     errorHandler?: () => void;
     externalAccessService?: TransformedAuthService;
     accessToken?: string;
@@ -207,7 +205,7 @@ function getPositionData({
   currentCanvas: TransformedCanvas;
   searchResults: SearchResults | null;
   canvases: TransformedCanvas[];
-  rotatedImages: RotatedImage[];
+  rotatedImages: CanvasRotatedImage[];
 }): OverlayPositionData[] {
   const searchHitsPositioningData = searchResults?.resources.map(resource => {
     // on: "https://wellcomelibrary.org/iiif/b30330002/canvas/c55#xywh=2301,662,157,47"
@@ -258,7 +256,7 @@ const ItemRenderer = memo(({ style, index, data }: ItemRendererProps) => {
   const currentCanvas = canvases[index];
   const { userIsStaffWithRestricted } = useUserContext();
   const isRestricted = currentCanvas.hasRestrictedImage;
-  const { searchResults, rotatedImages } = useContext(ItemViewerContext);
+  const { searchResults, rotatedImages } = useItemViewerContext();
   const [imageRect, setImageRect] = useState<DOMRect | undefined>();
   const [imageContainerRect, setImageContainerRect] = useState<
     DOMRect | undefined
@@ -409,7 +407,7 @@ const MainViewer: FunctionComponent = () => {
     setShowControls,
     errorHandler,
     accessToken,
-  } = useContext(ItemViewerContext);
+  } = useItemViewerContext();
   const { shouldScrollToCanvas, canvas } = query;
   const mainViewerRef = useRef<FixedSizeList>(null);
   const [newScrollOffset, setNewScrollOffset] = useState(0);
