+++ content/webapp/components/IIIFViewer/ZoomedImage.tsx
@@ -2,7 +2,6 @@ import openseadragon from 'openseadragon';
 import {
   FunctionComponent,
   MutableRefObject,
-  useContext,
   useEffect,
   useRef,
   useState,
@@ -14,7 +13,7 @@ import { DigitalLocation } from '@weco/common/model/catalogue';
 import { OptionalToUndefined } from '@weco/common/utils/utility-types';
 import Control from '@weco/common/views/components/Control';
 import Space from '@weco/common/views/components/styled/Space';
-import ItemViewerContext from '@weco/content/contexts/ItemViewerContext';
+import { useItemViewerContext } from '@weco/content/contexts/ItemViewerContext';
 import { convertRequestUriToInfoUri } from '@weco/content/utils/iiif/convert-iiif-uri';
 
 import { queryParamToArrayIndex } from '.';
@@ -52,8 +51,7 @@ type ZoomedImageProps = OptionalToUndefined<{
 const ZoomedImage: FunctionComponent<ZoomedImageProps> = ({
   iiifImageLocation,
 }) => {
-  const { transformedManifest, query, setShowZoomed } =
-    useContext(ItemViewerContext);
+  const { transformedManifest, query, setShowZoomed } = useItemViewerContext();
   const currentCanvas =
     transformedManifest?.canvases[queryParamToArrayIndex(query.canvas)];
   const mainImageService = {
