--- content/webapp/components/ImageGallery/README.md
@@ -1,5 +1,0 @@
-## Purpose
-To enable the display of a number of images without requiring a user to view them.
-
-
-[Edit this Readme on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/main/content/webapp/components/ImageGallery/README.md)
